CRUD Movie

+++ src/main/java/com/sba301/online_ticket_sales/dto/person/request/PersonUpdateRequest.java
@@ -15,6 +15,6 @@ public class PersonUpdateRequest {
     private Double height;
     private Occupation occupation;
     private String biography;
-    private Long countryId;
+    private Integer countryId;
     private List<String> images;
 }
